chore: bump go-waku

+++ eth-node/types/mailserver.go
@@ -35,9 +35,9 @@ type MessagesRequest struct {
 }
 
 type StoreRequestCursor struct {
-	Digest       []byte  `json:"digest"`
-	ReceiverTime float64 `json:"receiverTime"`
-	SenderTime   float64 `json:"senderTime"`
+	Digest       []byte `json:"digest"`
+	ReceiverTime int64  `json:"receiverTime"`
+	SenderTime   int64  `json:"senderTime"`
 }
 
 // SetDefaults sets the From and To defaults
